requirements.txt
django==3.0 
pillow
djangorestframework
requests
django-progressive-web-app
